dashboard/requirements.txt
pandas
altair
python-dotenv
streamlit
watchdog
boto3
